# docs/swagger.yaml
definitions:
...
  v1.ContentItemDetailResponseData:
    properties:
      audioProgressSeconds:
        type: integer
      availableText:
        type: string
      contentSafe:
        type: string
      contentType:
        example: text
        type: string
      descriptionSafe:
        type: string
      favorited:
        type: boolean
      feedId:
        type: integer
      fetchedAt:
        type: string
      id:
        type: integer
      markedLater:
        type: boolean
      processingStatus:
        example: unprocessed
        type: string
      publishedAt:
        type: string
      showNotesSafe:
        type: string
      title:
        type: string
    type: object
  v1.ContentItemListItemData:
    properties:
      audioProgressSeconds:
        type: integer
      availableText:
        type: string
      contentType:
        example: text
        type: string
      favorited:
        type: boolean
      feedId:
        type: integer
      fetchedAt:
        type: string
      id:
        type: integer
      markedLater:
        type: boolean
      processingStatus:
        example: unprocessed
        type: string
      publishedAt:
        type: string
      title:
        type: string
    type: object
...
  v1.GetContentItemResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/v1.ContentItemDetailResponseData'
      message:
        type: string
    type: object
...
  v1.ListContentItemsResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/v1.ListContentItemsResponseData'
      message:
        type: string
    type: object
  v1.ListContentItemsResponseData:
    properties:
      items:
        items:
          $ref: '#/definitions/v1.ContentItemListItemData'
        type: array
      page:
        $ref: '#/definitions/v1.PageMeta'
    type: object
...
  v1.PageMeta:
    properties:
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
...
  v1.UpdateContentItemAudioProgressRequest:
    properties:
      audioProgressSeconds:
        type: integer
    required:
    - audioProgressSeconds
    type: object
  v1.UpdateContentItemMarkRequest:
    properties:
      marked:
        type: boolean
    required:
    - marked
    type: object
  v1.UpdateContentItemProcessingStatusRequest:
    properties:
      processingStatus:
        type: string
    required:
    - processingStatus
    type: object
host: localhost:8000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server celler server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Nunu Example API
  version: 1.0.0
paths:
  /content-items:
    get:
      description: 统一内容列表查询入口，支持关键词搜索、内容类型、处理状态、内容标记和订阅源单值 AND 过滤
      parameters:
      - description: 'content type: text/audio'
        in: query
        name: content_type
        type: string
      - description: 'processing status: unprocessed/completed'
        in: query
        name: processing_status
        type: string
      - description: 'content mark: later/favorite'
        in: query
        name: mark
        type: string
      - description: feed id
        in: query
        name: feed_id
        type: integer
      - description: tag id
        in: query
        name: tag_id
        type: integer
      - description: search keyword, max 128 characters
        in: query
        name: keyword
        type: string
      - description: page
        in: query
        name: page
        type: integer
      - description: page size
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.ListContentItemsResponse'
      security:
      - Bearer: []
      summary: 查询统一内容条目列表
      tags:
      - 内容条目模块
  /content-items/{id}:
    get:
      description: 返回单条内容条目的详情与安全渲染字段
      parameters:
      - description: content item id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.GetContentItemResponse'
      security:
      - Bearer: []
      summary: 查询内容条目详情
      tags:
      - 内容条目模块
  /content-items/{id}/audio-progress:
    put:
      consumes:
      - application/json
      description: 仅持久化音频型内容条目的播放位置，不持久化文本阅读位置，也不改变处理状态。
      parameters:
      - description: content item id
        in: path
        name: id
        required: true
        type: integer
      - description: audio progress seconds
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v1.UpdateContentItemAudioProgressRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.GetContentItemResponse'
      security:
      - Bearer: []
      summary: 更新音频播放进度
      tags:
      - 内容条目模块
  /content-items/{id}/marks/{mark}:
    put:
      consumes:
      - application/json
      description: 独立设置或取消稍后处理、收藏标记，不改变处理状态或消费进度。
      parameters:
      - description: content item id
        in: path
        name: id
        required: true
        type: integer
      - description: 'content mark: later/favorite'
        in: path
        name: mark
        required: true
        type: string
      - description: content mark state
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v1.UpdateContentItemMarkRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.GetContentItemResponse'
      security:
      - Bearer: []
      summary: 更新内容条目标记
      tags:
      - 内容条目模块
  /content-items/{id}/processing-status:
    put:
      consumes:
      - application/json
      description: 手动在未处理和已完成之间切换内容条目的处理状态，不改变内容标记或消费进度。
      parameters:
      - description: content item id
        in: path
        name: id
        required: true
        type: integer
      - description: processing status
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v1.UpdateContentItemProcessingStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.GetContentItemResponse'
      security:
      - Bearer: []
      summary: 更新内容条目处理状态
      tags:
      - 内容条目模块
...
  /content-views/completed:
    get:
      description: Completed view presets processing_status=completed and accepts
        additional single-value filters.
      parameters:
      - description: 'content type: text/audio'
        in: query
        name: content_type
        type: string
      - description: 'content mark: later/favorite'
        in: query
        name: mark
        type: string
      - description: feed id
        in: query
        name: feed_id
        type: integer
      - description: tag id
        in: query
        name: tag_id
        type: integer
      - description: search keyword, max 128 characters
        in: query
        name: keyword
        type: string
      - description: page
        in: query
        name: page
        type: integer
      - description: page size
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.ListContentItemsResponse'
      security:
      - Bearer: []
      summary: List completed content items
      tags:
      - Content item module
  /content-views/favorite:
    get:
      description: Favorite view presets mark=favorite and accepts additional single-value
        filters.
      parameters:
      - description: 'content type: text/audio'
        in: query
        name: content_type
        type: string
      - description: 'processing status: unprocessed/completed'
        in: query
        name: processing_status
        type: string
      - description: feed id
        in: query
        name: feed_id
        type: integer
      - description: tag id
        in: query
        name: tag_id
        type: integer
      - description: search keyword, max 128 characters
        in: query
        name: keyword
        type: string
      - description: page
        in: query
        name: page
        type: integer
      - description: page size
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.ListContentItemsResponse'
      security:
      - Bearer: []
      summary: List favorite content items
      tags:
      - Content item module
  /content-views/inbox:
    get:
      description: Inbox view presets processing_status=unprocessed and accepts additional
        single-value filters.
      parameters:
      - description: 'content type: text/audio'
        in: query
        name: content_type
        type: string
      - description: 'content mark: later/favorite'
        in: query
        name: mark
        type: string
      - description: feed id
        in: query
        name: feed_id
        type: integer
      - description: tag id
        in: query
        name: tag_id
        type: integer
      - description: search keyword, max 128 characters
        in: query
        name: keyword
        type: string
      - description: page
        in: query
        name: page
        type: integer
      - description: page size
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.ListContentItemsResponse'
      security:
      - Bearer: []
      summary: List inbox content items
      tags:
      - Content item module
  /content-views/later:
    get:
      description: Later view presets mark=later and accepts additional single-value
        filters.
      parameters:
      - description: 'content type: text/audio'
        in: query
        name: content_type
        type: string
      - description: 'processing status: unprocessed/completed'
        in: query
        name: processing_status
        type: string
      - description: feed id
        in: query
        name: feed_id
        type: integer
      - description: tag id
        in: query
        name: tag_id
        type: integer
      - description: search keyword, max 128 characters
        in: query
        name: keyword
        type: string
      - description: page
        in: query
        name: page
        type: integer
      - description: page size
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.ListContentItemsResponse'
      security:
      - Bearer: []
      summary: List later content items
      tags:
      - Content item module
...
  /feeds/{id}/content-items:
    get:
      description: Feed detail view presets feed_id from the path and accepts additional
        single-value filters.
      parameters:
      - description: feed id
        in: path
        name: id
        required: true
        type: integer
      - description: 'content type: text/audio'
        in: query
        name: content_type
        type: string
      - description: 'processing status: unprocessed/completed'
        in: query
        name: processing_status
        type: string
      - description: 'content mark: later/favorite'
        in: query
        name: mark
        type: string
      - description: tag id
        in: query
        name: tag_id
        type: integer
      - description: search keyword, max 128 characters
        in: query
        name: keyword
        type: string
      - description: page
        in: query
        name: page
        type: integer
      - description: page size
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.ListContentItemsResponse'
      security:
      - Bearer: []
      summary: List feed content items
      tags:
      - Content item module
...
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"